+++ CHANGELOG.md
@@ -6,6 +6,7 @@ TBD
 
 - Added max_concurrent support for config.yml under `plugin.local.max_concurrent`
 - Added max concurrent support for beer.conf under `MAX_CONCURRENT`
+- Added feature to cancel an ongoing request or delete a completed request
 - Fixed API swagger docs for Job execution
 - Fixed issue where using local systems with a Namespace not the Garden name would not be treated as a local
   Request for internal features