+++ CHANGELOG.md
@@ -11,6 +11,7 @@ TBD
 - Fixed issue where using local systems with a Namespace not the Garden name would not be treated as a local
   Request for internal features
 - Added config option to import scheduled jobs file at startup
+- Added config option for child garden configuration file to share all users with child garden instead of filtering
 
 # 3.27.2