+++ CHANGELOG.md
@@ -13,6 +13,7 @@ TBD
 - Added config option to import scheduled jobs file at startup
 - Added support to configure `alias_user_mappings` in the user configuration file
 - Added config option for child garden configuration file to share all users with child garden instead of filtering
+- Added System Admin feature to Cancel Non-Completed Requests in bulk, instead of Deleting
 
 # 3.27.2