+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 TBD
 
+- Added support for group/groups in beer.conf and group environment variable
 - Added max_concurrent support for config.yml under `plugin.local.max_concurrent`
 - Added max concurrent support for beer.conf under `MAX_CONCURRENT`
 - Added feature to cancel an ongoing request or delete a completed request