+++ CHANGELOG.md
@@ -14,6 +14,7 @@ TBD
 - Added config option to import scheduled jobs file at startup
 - Added support to configure `alias_user_mappings` in the user configuration file
 - Added config option for child garden configuration file to share all users with child garden instead of filtering
+- Added support for pruning completed orphaned child requests
 
 # 3.27.2