+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 TBD
 
+- Updated create request page warning to identify required systems not in running status.
 - Added support to prune legacy request with no command type
 - Added support for group/groups in beer.conf and group environment variable
 - Added max_concurrent support for config.yml under `plugin.local.max_concurrent`